+++ buession-security-captcha/src/main/java/com/buession/security/captcha/aliyun/AliYunCaptchaClient.java
@@ -24,8 +24,8 @@
  */
 package com.buession.security.captcha.aliyun;
 
-import com.buession.core.builder.MapBuilder;
 import com.buession.core.utils.Assert;
+import com.buession.core.validator.Validate;
 import com.buession.httpclient.HttpClient;
 import com.buession.httpclient.core.Response;
 import com.buession.lang.Status;
@@ -33,14 +33,13 @@ import com.buession.security.captcha.AbstractCaptchaClient;
 import com.buession.security.captcha.core.CaptchaException;
 import com.buession.security.captcha.core.Manufacturer;
 import com.buession.security.captcha.core.RequestData;
+import com.buession.security.captcha.core.RequiredParameterCaptchaException;
 import com.buession.security.captcha.utils.ObjectMapperUtils;
-import com.buession.security.mcrypt.Algo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.Objects;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * 阿里云行为验证 Client，文档：<a href="https://help.aliyun.com/document_detail/28309.html" target="_blank">https://help.aliyun.com/document_detail/28309.html</a>
@@ -50,21 +49,20 @@ import java.util.Objects;
  */
 public class AliYunCaptchaClient extends AbstractCaptchaClient {
 
-	public final static String ACTION = "AuthenticateSig";
-
-	public final static String FORMAT = "JSON";
-
-	public final static String SIGNATURE_VERSION = "1.0";
-
 	/**
-	 * 密钥 ID
+	 * AccessKey ID
 	 */
 	private final String accessKeyId;
 
 	/**
-	 * 服务使用的 App Key
+	 * AccessKey Secret
 	 */
 	private final String accessKeySecret;
+
+	/**
+	 * 服务使用的 App Key
+	 */
+	private final String appKey;
 
 	/**
 	 * 端点
@@ -77,29 +75,36 @@ public class AliYunCaptchaClient extends AbstractCaptchaClient {
 	 * 构造函数
 	 *
 	 * @param accessKeyId
-	 * 		密钥 ID
+	 * 		AccessKey ID
 	 * @param accessKeySecret
+	 * 		AccessKey Secret
+	 * @param appKey
 	 * 		服务使用的 App Key
 	 */
-	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret){
+	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret, final String appKey){
 		Assert.isBlank(accessKeyId, "AccessKeyId cloud not be empty or null");
 		Assert.isBlank(accessKeySecret, "AccessKeySecret cloud not be empty or null");
+		Assert.isBlank(appKey, "AppKey cloud not be empty or null");
 		this.accessKeyId = accessKeyId;
 		this.accessKeySecret = accessKeySecret;
+		this.appKey = appKey;
 	}
 
 	/**
 	 * 构造函数
 	 *
 	 * @param accessKeyId
-	 * 		密钥 ID
+	 * 		AccessKey ID
 	 * @param accessKeySecret
-	 * 		阿里云验证码配置对应的 App Key
+	 * 		AccessKey Secret
+	 * @param appKey
+	 * 		服务使用的 App Key
 	 * @param httpClient
 	 *        {@link HttpClient} 实例
 	 */
-	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret, final HttpClient httpClient){
-		this(accessKeyId, accessKeySecret);
+	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret, final String appKey,
+							   final HttpClient httpClient){
+		this(accessKeyId, accessKeySecret, appKey);
 		setHttpClient(httpClient);
 	}
 
@@ -107,14 +112,17 @@ public class AliYunCaptchaClient extends AbstractCaptchaClient {
 	 * 构造函数
 	 *
 	 * @param accessKeyId
-	 * 		密钥 ID
+	 * 		AccessKey ID
 	 * @param accessKeySecret
-	 * 		阿里云验证码配置对应的 App Key
+	 * 		AccessKey Secret
+	 * @param appKey
+	 * 		服务使用的 App Key
 	 * @param regionId
 	 * 		区域 ID
 	 */
-	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret, final String regionId){
-		this(accessKeyId, accessKeySecret);
+	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret, final String appKey,
+							   final String regionId){
+		this(accessKeyId, accessKeySecret, appKey);
 
 		Assert.isBlank(regionId, "RegionId cloud not be empty or null");
 
@@ -125,17 +133,19 @@ public class AliYunCaptchaClient extends AbstractCaptchaClient {
 	 * 构造函数
 	 *
 	 * @param accessKeyId
-	 * 		密钥 ID
+	 * 		AccessKey ID
 	 * @param accessKeySecret
-	 * 		阿里云验证码配置对应的 App Key
+	 * 		AccessKey Secret
+	 * @param appKey
+	 * 		服务使用的 App Key
 	 * @param regionId
 	 * 		区域 ID
 	 * @param httpClient
 	 * 		Http Client
 	 */
-	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret, final String regionId,
-							   final HttpClient httpClient){
-		this(accessKeyId, accessKeySecret, regionId);
+	public AliYunCaptchaClient(final String accessKeyId, final String accessKeySecret, final String appKey,
+							   final String regionId, final HttpClient httpClient){
+		this(accessKeyId, accessKeySecret, appKey, regionId);
 		setHttpClient(httpClient);
 	}
 
@@ -146,72 +156,42 @@ public class AliYunCaptchaClient extends AbstractCaptchaClient {
 		}
 
 		AliYunRequestData aliYunRequestData = (AliYunRequestData) requestData;
-
-		SimpleDateFormat sdf = new SimpleDateFormat("YYYY-MM-DDThh:mm:ssZ");
-		MapBuilder<String, Object> parametersBuilder = MapBuilder.<String, Object>create()
-				.put("Action", ACTION).put("Format", FORMAT).put("Version", getVersion())
-				.put("SignatureMethod", Algo.HMAC_SHA1.getName()).put("SignatureNonce", randomStr())
-				.put("SignatureVersion", SIGNATURE_VERSION)
-				.put("AccessKeyId", accessKeyId).put("AppKey", accessKeySecret)
-				.put("Timestamp", sdf.format(new Date()));
-
-		//sb.append("&Signature=").append(getVersion());
-
-		if(aliYunRequestData.getToken() != null){
-			parametersBuilder.put("Token", aliYunRequestData.getToken());
+		if(checkParam(aliYunRequestData) == false){
+			return Status.FAILURE;
 		}
 
-		if(aliYunRequestData.getSig() != null){
-			parametersBuilder.put("Sig", aliYunRequestData.getSig());
-		}
-
-		if(aliYunRequestData.getSessionId() != null){
-			parametersBuilder.put("SessionId", aliYunRequestData.getSessionId());
-		}
-
-		if(aliYunRequestData.getScene() != null){
-			parametersBuilder.put("Scene", aliYunRequestData.getScene());
-		}
-
-		if(aliYunRequestData.getClientIp() != null){
-			parametersBuilder.put("RemoteIp", aliYunRequestData.getClientIp());
-		}
-
-		if(aliYunRequestData.getSourceIp() != null){
-			parametersBuilder.put("SourceIp", aliYunRequestData.getSourceIp());
-		}
+		AliyunParametersBuilder parametersBuilder = new AliyunParametersBuilder(accessKeyId,
+				accessKeySecret, appKey, this);
+		Map<String, Object> parameters = new HashMap<>(parametersBuilder.build(aliYunRequestData));
 
 		if(logger.isDebugEnabled()){
-			logger.debug("二次验证, parameters：{}.", parametersBuilder.build());
+			logger.debug("二次验证, parameters：{}.", parameters);
 		}
 
 		Response response;
 		try{
-			response = getHttpClient().get(endpoint, parametersBuilder.build());
+			response = getHttpClient().get(endpoint, parameters, getHeaders());
 
 			if(logger.isInfoEnabled()){
 				logger.info("二次验证 response: {}", response);
 			}
 
-			AliyunEnhencedResult result = ObjectMapperUtils.createObjectMapper().readValue(response.getBody(),
-					AliyunEnhencedResult.class);
+			if(response.isSuccessful()){
+				AliyunEnhencedResult result = ObjectMapperUtils.createObjectMapper().readValue(response.getBody(),
+						AliyunEnhencedResult.class);
 
-			if(Objects.equals(result.getSuccess(), Boolean.TRUE)){
-				if(result.getData() == null){
-					throw new CaptchaException("Validate failure");
-				}
-
-				if(Objects.equals(result.getData().getCode(), "100")){
+				if(result.getCode() == 100){
 					return Status.SUCCESS;
 				}else{
-					throw new CaptchaException("Validate failure(code: " + result.getData().getCode() + ")");
+					throw new CaptchaException(
+							"Validate failure: " + result.getMsg() + "(code: " + result.getCode() + ")");
 				}
 			}else{
-				if(Objects.equals(result.getCode(), "400")){
+				if(response.getStatusCode() == 400){
 					throw new CaptchaException("Request parameter error");
-				}else{
-					throw new CaptchaException("Aliyun afs server system error");
 				}
+
+				return Status.FAILURE;
 			}
 		}catch(Exception e){
 			logger.error("二次验证失败: {}", e.getMessage());
@@ -236,6 +216,35 @@ public class AliYunCaptchaClient extends AbstractCaptchaClient {
 	 */
 	public String getEndpoint(){
 		return endpoint;
+	}
+
+	/**
+	 * 检查客户端的请求是否合法，只要有一个为空，则判断不合法
+	 *
+	 * @param requestData
+	 *        {@link AliYunRequestData}
+	 *
+	 * @return 检测结果
+	 */
+	private static boolean checkParam(final AliYunRequestData requestData)
+			throws RequiredParameterCaptchaException{
+		if(Validate.hasText(requestData.getToken()) == false){
+			throw new RequiredParameterCaptchaException("Token");
+		}
+
+		if(Validate.hasText(requestData.getSig()) == false){
+			throw new RequiredParameterCaptchaException("Sig");
+		}
+
+		if(Validate.hasText(requestData.getSessionId()) == false){
+			throw new RequiredParameterCaptchaException("SessionId");
+		}
+
+		if(Validate.hasText(requestData.getScene()) == false){
+			throw new RequiredParameterCaptchaException("Scene");
+		}
+
+		return true;
 	}
 
 }