+++ buession-security-captcha/src/main/java/com/buession/security/captcha/core/Manufacturer.java
@@ -43,11 +43,6 @@ public enum Manufacturer {
 	GEETEST,
 
 	/**
-	 * 网易
-	 */
-	NETEASE,
-
-	/**
 	 * 腾讯
 	 */
 	TENCENT